+++ app/forms/history/[id]/page.tsx
@@ -1,12 +1,12 @@
 import HistoryPanel from './panel';
 import { getRecordHistory, loadAllRecords } from '../../../../components/chain_utils';
 
-export async function generateStaticParams(){
-  const records = await loadAllRecords();
-  return records.map(record => ({
-    id: record.id,
-  }));
-}
+// export async function generateStaticParams(){
+//   const records = await loadAllRecords();
+//   return records.map(record => ({
+//     id: record.id,
+//   }));
+// }
 
 export default async function Page({ params }){
   const recordID: string = params.id;